refactor(attention): the incident-grouping cases are their own file
worklist_test.go crossed the 1000-line ceiling. The cases about when many
broken things are ONE broken thing — what folds, what refuses to, and what must
never be folded because a customer is on the other end of it — are one subject,
and the file's other cases are about ranking a single row.

The migration is re-stamped above the ones that landed on main meanwhile.
